requirements.txt
discord~=2.3.2
python-dotenv~=1.0.1
requests~=2.31.0
mihomo~=2.0.2